# Avoid allocating the CandidateSet when there's a single match