--- a/src/Http/Routing/src/EndpointSelectorContext.cs
+++ b/src/Http/Routing/src/EndpointSelectorContext.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
@@ -24,7 +24,7 @@ namespace Microsoft.AspNetCore.Routing
         /// </summary>
         public RouteValueDictionary RouteValues
         {
-            get => _routeValues;
+            get => _routeValues ?? (_routeValues = new RouteValueDictionary());
             set
             {
                 _routeValues = value;